+++ lucasalary/lib/luca_salary/payment.rb
@@ -99,8 +99,8 @@ module LucaSalary
       [].tap do |res|
         {}.tap do |item|
           item['date'] = "#{accrual_date.year}-#{accrual_date.month}-#{accrual_date.day}"
-          item['debit'] = h[:debit].map { |k, v| { 'label' => k, 'value' => v } }
-          item['credit'] = h[:credit].map { |k, v| { 'label' => k, 'value' => v } }
+          item['debit'] = h[:debit].map { |k, v| { 'label' => k, 'amount' => v } }
+          item['credit'] = h[:credit].map { |k, v| { 'label' => k, 'amount' => v } }
           item['x-editor'] = 'LucaSalary'
           res << item
         end